Server/requirements.txt
# Core dependencies
fastapi[all]>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
openai>=1.0
python-dotenv>=1.0
tenacity>=8.0
python-dateutil>=2.8
PyYAML>=6.0
python-Levenshtein>=0.21.0
